# swagger-new.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "IronLogic Fitness Training API",
    "description": "API for managing fitness training data including clients, activities, schedules, and benchmarks",
    "version": "1.0.0",
    "contact": {
      "name": "IronLogic API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/other-benchmarks": {
      "get": {
        "tags": ["Benchmarks"],
        "summary": "Get all other benchmarks",
        "description": "Retrieve a list of all other benchmarks",
        "responses": {
          "200": {
            "description": "List of other benchmarks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "count": {
                      "type": "integer"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/OtherBenchmark"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "post": {
        "tags": ["Benchmarks"],
        "summary": "Create a new other benchmark",
        "description": "Create a new other benchmark",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200,
                    "description": "Benchmark name"
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Benchmark notes"
                  },
                  "measurementNotes": {
                    "type": "string",
                    "maxLength": 1000,
                    "description": "Measurement details"
                  },
                  "value": {
                    "type": "number",
                    "minimum": 0,
                    "description": "Numeric value"
                  },
                  "unit": {
                    "type": "string",
                    "maxLength": 50,
                    "description": "Unit of measurement"
                  },
                  "benchmarkTemplateId": {
                    "type": "string",
                    "description": "Reference to benchmark template"
                  }
                }
              },
              "example": {
                "name": "5K Run Time",
                "notes": "Personal best 5K time",
                "measurementNotes": "Measured on track",
                "value": 22.5,
                "unit": "minutes",
                "benchmarkTemplateId": "template456"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Other benchmark created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/OtherBenchmark"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
    "/other-benchmarks/{id}": {
      "get": {
        "tags": ["Benchmarks"],
        "summary": "Get other benchmark by ID",
        "description": "Get a specific other benchmark by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Other benchmark ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Other benchmark found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/OtherBenchmark"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Other benchmark not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "put": {
        "tags": ["Benchmarks"],
        "summary": "Update other benchmark",
        "description": "Update an other benchmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Other benchmark ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "measurementNotes": {
                    "type": "string",
                    "maxLength": 1000
                  },
                  "value": {
                    "type": "number",
                    "minimum": 0
                  },
                  "unit": {
                    "type": "string",
                    "maxLength": 50
                  },
                  "benchmarkTemplateId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Other benchmark updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/OtherBenchmark"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Other benchmark not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      },
      "delete": {
        "tags": ["Benchmarks"],
        "summary": "Delete other benchmark",
        "description": "Delete an other benchmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Other benchmark ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Other benchmark deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Other benchmark not found"
          },
          "401": {
            "description": "Authentication required"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OtherBenchmark": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Benchmark ID"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Benchmark name"
          },
          "notes": {
            "type": "string",
            "maxLength": 1000,
            "description": "Benchmark notes"
          },
          "measurementNotes": {
            "type": "string",
            "maxLength": 1000,
            "description": "Measurement details"
          },
          "value": {
            "type": "number",
            "minimum": 0,
            "description": "Numeric value"
          },
          "unit": {
            "type": "string",
            "maxLength": 50,
            "description": "Unit of measurement"
          },
          "benchmarkTemplateId": {
            "type": "string",
            "description": "Reference to benchmark template"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["name"]
      },
...
    }
  }
} 